Rename OpenSSLSerializationBackend
