+++ pytest.ini
@@ -8,4 +8,5 @@ markers =
     hmac: this test requires a backend providing HMACBackend
     pbkdf2hmac: this test requires a backend providing PBKDF2HMACBackend
     rsa: this test requires a backend providing RSABackend
+    traditional_openssl_serialization: this test requires a backend providing TraditionalOpenSSLSerializationBackend
     supported: parametrized test requiring only_if and skip_message
